requirements.txt
emoji~=2.12.1
Pillow~=10.4.0
PySide6==6.7.2
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
Requests~=2.32.3
